[NXP toup] rpmsg-lite: make timeout value to be aarch64 compatible
Some time related macro such as RL_BLOCK, RL_TRUE and RL_FALSE
are defined as "UL", so timeout value should be 64bit for aarch64
and 32bit for aarch32.

+++ lib/rpmsg_lite/porting/environment/rpmsg_env_bm.c
@@ -75,7 +75,7 @@ static struct isr_info isr_table[ISR_COUNT];
  * Busy loop implementation for BM.
  *
  */
-uint32_t env_wait_for_link_up(volatile uint32_t *link_state, uint32_t link_id, uint32_t timeout_ms)
+uint32_t env_wait_for_link_up(volatile uint32_t *link_state, uint32_t link_id, uintptr_t timeout_ms)
 {
     uint32_t tick_count = 0U;
     uint32_t tick_temp;
